+++ src/app/models/Sparql.ts
@@ -45,3 +45,8 @@ export class QueryResultBinding {
     datatype?: string;
     'xml:lang': string;
 }
+
+export class FederatedEndpointSuggestion {
+    endpointURL: string;
+    endpointLabel: string;
+}